+++ cmd/tempo/app/modules.go
@@ -102,6 +102,8 @@ func (t *App) initOverrides() (services.Service, error) {
 	}
 	t.overrides = overrides
 
+	t.Server.HTTP.Handle("/runtime_config", overrides.Handler())
+
 	return t.overrides, nil
 }
 
@@ -282,8 +284,8 @@ func (t *App) setupModuleManager() error {
 
 	deps := map[string][]string{
 		// Server:       nil,
-		// Overrides:    nil,
 		// Store:        nil,
+		Overrides:     {Server},
 		MemberlistKV:  {Server},
 		QueryFrontend: {Server},
 		Ring:          {Server, MemberlistKV},